backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# PDF generation
weasyprint==62.3

# Encryption (IBAN)
cryptography==43.0.3

# Utilities
python-dateutil==2.9.0
pytz==2024.2
python-dotenv==1.0.1

# Dev / tests
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
